Refactored the CSS and imprved code readbility
