+++ frontend/components/stops/StopProbeHeader/index.tsx
@@ -3,7 +3,7 @@
 import { Button } from '@mantine/core';
 import { useTranslations } from 'next-intl';
 
-import styles from '../StopProbeSection/styles.module.css';
+import styles from './styles.module.css';
 
 /* * */
