+++ frontend/components/stops/StopProbeQuestion1/styles.module.css
@@ -1,0 +1,36 @@
+.fullWidth {
+    width: 100%;
+  }
+
+  /* PROBE TITLE */
+
+.probeTitle{
+    display: flex;
+    align-items: center;
+    font-size: var(--font-size-title);
+    font-weight: var(--font-weight-title);
+    color: var(--color-system-text-100);
+  }
+  
+  /* * */
+
+  /* POSSIBLE ANSWERS */
+.possibleAnswersContainer {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 10px;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+    padding: 25px 0 0;
+  }
+  
+  .possibleAnswersContainer span{
+  pointer-events: none;
+  }
+  
+  .selected{
+    border: 2px solid var(--color-brand);
+  }
+  
+  /* * */
