+++ frontend/components/stops/StopProbeQuestion3/index.tsx
@@ -3,7 +3,7 @@ import { Radio } from '@mantine/core';
 import { useTranslations } from 'next-intl';
 import { useState } from 'react';
 
-import styles from '../StopProbeSection/styles.module.css';
+import styles from './styles.module.css';
 
 interface Props {
 	question3Answer: string
